Fix Camel case bug when a property is of type Enum or another Class
