+++ src/TypeScriptDefinitionGenerator.csproj
@@ -56,6 +56,7 @@
             <DependentUpon>DtsPackage.vsct</DependentUpon>
         </Compile>
         <Compile Include="Helpers\Telemetry.cs" />
+        <Compile Include="Helpers\Utility.cs" />
         <Compile Include="Helpers\VSHelpers.cs" />
         <Compile Include="Options.cs">
             <SubType>Component</SubType>
